put heuristics in separate modules
